chore: move domain schemas to separate module
